Test database creation on docker container start.
A new database is created with the same credentials on restart.
The `build_sample_db.sh` script get's copied to container and run.
This script then waits for the mysql database to start and runs
a query to generate a new database called `test` with a table called
`users` containing one default user called `admin`.